% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fetch.R
\name{cmip_fetch}
\alias{cmip_fetch}
\title{Fetch data}
\usage{
cmip_fetch(route, overwrite = FALSE)
}
\arguments{
\item{route}{Route to a folder, a character string}

\item{overwrite}{(logical) Whether to overwrite files if they already
exist on your machine. Default: \code{FALSE}}
}
\value{
A character vector of full file paths. A print method makes a tidy
return object
}
\description{
Fetch data
}
\details{
Note that data is not read into R as data can be very large.
See \code{\link{cmip_read}}
}
\examples{
\dontrun{
key <- "bcsd/yearly/cnrm_cm3.1/cnrm_cm3.1.sresa1b.monthly.Prcp.2034.nc"
(res <- cmip_fetch(key))
cmip_read(res)
}
}

