% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip_read.R
\name{cmip_read}
\alias{cmip_read}
\title{Reaa data into R}
\usage{
cmip_read(x)
}
\arguments{
\item{x}{A \code{ccafs_files} object, the output from a call to
\code{\link{cmip_fetch}}}
}
\value{
\code{RasterLayer} or \code{RasterBrick} class object
}
\description{
Reaa data into R
}
\examples{
\dontrun{
key <- "bcsd/yearly/cnrm_cm3.1/cnrm_cm3.1.sresa1b.monthly.Prcp.2035.nc"
res <- cmip_fetch(key)

# a single file
cmip_read(res)

# character path input
## you can also pass in a path to a file(s)
cmip_read(res[1])

# plot data
plot(cmip_read(res))
}
}
