% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-package.R
\docType{package}
\name{cmipr-package}
\alias{cmipr-package}
\alias{cmipr}
\title{cmipr}
\description{
R client for CMIP (Coupled Model Intercomparison Project) data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
