% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files/folders}
\usage{
list_files(route = NULL)
}
\arguments{
\item{route}{Route to a FTP folder on the CMIP site}
}
\value{
A \code{tibble}
}
\description{
List files/folders
}
\examples{
\dontrun{
list_files()
list_files('bcsd')
list_files('bcsd/yearly')
list_files('bcsd/yearly/cccma_cgcm3_1.1')
list_files('cmip5')
list_files('cmip5/bcsd')
list_files('cmip5/bcsd/regrid')
list_files('cmip5/bcsd/regrid/ccsm4')
list_files('cmip5/bcsd/regrid/ccsm4/rcp26')
list_files('cmip5/bcsd/regrid/ccsm4/rcp26/mon')
list_files('cmip5/bcsd/regrid/ccsm4/rcp26/mon/r3i1p1')
list_files('cmip5/bcsd/regrid/ccsm4/rcp26/mon/r3i1p1/tas')
list_files('wwcra/monthly/riog/')
list_files('wwcra/monthly/riog/sresb1.mpi_echam5.3/tmax')
}
}

